Split test code from lib code for dodgy_dd.
